Improved logging with progress bars and in CI + Corrected Docker configuration for CIv2
* Added a way for Ruby actions to populate action's stdout and stderr
* Better logging
* Limit error messages length in Confluence as now logging add much more info into errors
* Typo
* Improved logging verbosity + Handled progress bar progression within other logs
* Cleaned logs
* Make idem_potence tests use ciadm after the first deployment
* Fixed wrong logging output with progress bar
* Added privileges to Docker containers
* Don't fail when connection was lost at the time of the close

Merge pull request #38 in PROJECThybrid_platforms_conductor from ~USERNME/hybrid_platforms_conductor:docker_in_ci_tests to master

Squashed commit of the following:

commit 42c8373e607e520af3243b79e696f28a392c352f
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Wed Mar 27 10:35:34 2019 +0100

    Remove containers unless reused to save place on Docker

commit c4f787a780b6951a53d050995ba810b4f478cc80
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Tue Mar 26 15:38:46 2019 +0100

    Corrected obsolete interfaces

commit a1bcf2878ff54d0633c4f8c5b23859525d84cc2c
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Tue Mar 26 13:50:45 2019 +0100

    Wait for SSH connections to be running

commit aaac2abd6ace632c89f9aea0f0aa9fd5a05271cf
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Thu Mar 14 14:44:39 2019 +0100

    Make sure connection test does not fail because of host key checking

commit 6fc997dc8c9199c7b850d51c98e169b680d91740
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Wed Mar 13 19:42:24 2019 +0100

    Wrong commit

commit 63ebbec3b756afb2553ded95e0b4bb425e92b97a
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Wed Mar 13 19:15:43 2019 +0100

    Don't fail when connection was lost at the time of the close

commit 1e2d29d79719c815ececf23c724fb75e6f707bde
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Wed Mar 13 17:57:57 2019 +0100

    Added privileges to Docker containers

commit 1f2708ce2f085c990eac3076e96074de56c168bc
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Wed Mar 13 16:21:02 2019 +0100

    Fixed wrong logging output with progress bar

commit b227f85a2cfec2850ed5b60efa1699385db1c3c3
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Tue Mar 12 20:58:58 2019 +0100

    Make idem_potence tests use ciadm after the first deployment

commit 40e65ee3a964e8d23f2eddde21296af8bfd49128
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Tue Mar 12 13:41:54 2019 +0100

    Cleaned logs

commit 0c4cd4b8b8d4b794199e82568d65fe7f48acb18a
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Tue Mar 12 13:03:03 2019 +0100

    Improved logging verbosity + Handled progress bar progression within other logs

commit b52e2faa814a6f1294b893d5291f665716a316d0
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Mon Mar 11 13:43:51 2019 +0100

    Typo

commit 3b697c2cdc98a0b08c21c64c668b3134c7a3fb47
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Mon Mar 11 13:30:45 2019 +0100

    Limit error messages length in Confluence as now logging add much more info into errors

commit d842560005475d585342f6d9f50ada184857ed81
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Mon Mar 11 13:26:13 2019 +0100

    Better logging

commit 07997634830e397f7efd6299ed42ed1de434cebf
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Mon Mar 11 13:13:43 2019 +0100

    Added a way for Ruby actions to populate action's stdout and stderr

commit 38794d3b83aa1956f20d96c7a13179ad9ac891bd
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Fri Mar 8 17:21:46 2019 +0100

    Better error logging

commit bc2d6990f5274211af7f23d0cb6aa0b1bb2c46c7
Author: Muriel Salvan <muriel@x-aeon.com>
Date:   Fri Mar 8 17:05:01 2019 +0100

    Debug parallelism

Commit comment amended by correctBitbucketSquashedMergeComment

+++ lib/hybrid_platforms_conductor/ssh_executor.rb
@@ -1,7 +1,6 @@
 require 'logger'
 require 'fileutils'
 require 'tmpdir'
-require 'tempfile'
 require 'hybrid_platforms_conductor/nodes_handler'
 require 'hybrid_platforms_conductor/cmd_runner'
 require 'hybrid_platforms_conductor/logger_helpers'
@@ -63,6 +62,8 @@ module HybridPlatformsConductor
       @ssh_env = {}
       @max_threads = 16
       @dry_run = false
+      @use_control_master = true
+      @strict_host_key_checking = true
       @override_connections = {}
       @passwords = {}
       @gateways_conf = ENV['ti_gateways_conf'].nil? ? :munich : ENV['ti_gateways_conf'].to_sym
@@ -99,6 +100,7 @@ module HybridPlatformsConductor
       out 'SSH executor configuration used:'
       out " * User: #{@ssh_user_name}"
       out " * Dry run: #{@dry_run}"
+      out " * Use SSH control master: #{@use_control_master}"
       out " * Max threads used: #{@max_threads}"
       out " * Gateways configuration: #{@gateways_conf}"
       out " * Gateway user: #{@gateway_user}"
@@ -174,9 +176,15 @@ module HybridPlatformsConductor
       options_parser.on('-g', '--gateway-user USER_NAME', "Name of the gateway user to be used by the XAE gateways (defaults to #{@gateway_user})") do |user_name|
         @gateway_user = user_name
       end
+      options_parser.on('-j', '--no-ssh-control-master', 'If used, don\'t create SSH control masters for connections.') do
+        @use_control_master = false
+      end
       options_parser.on('-m', '--max-threads NBR', "Set the number of threads to use for concurrent queries (defaults to #{@max_threads})") do |nbr_threads|
         @max_threads = nbr_threads.to_i
       end if parallel
+      options_parser.on('-q', '--no-ssh-host-key-checking', 'If used, don\'t check for SSH host keys.') do
+        @strict_host_key_checking = false
+      end
       options_parser.on('-s', '--show-commands', 'Display the SSH commands that would be run instead of running them') do
         self.dry_run = true
       end
@@ -295,6 +303,7 @@ Host *
   User #{@ssh_user_name}
   # Default control socket path to be used when multiplexing SSH connections
   ControlPath /tmp/hpc_ssh_executor_mux_%h_%p_%r
+  #{@strict_host_key_checking ? '' : 'StrictHostKeyChecking no'}
   #{known_hosts_file.nil? ? '' : "UserKnownHostsFile #{known_hosts_file}"}
   #{open_ssh_major_version >= 7 ? 'PubkeyAcceptedKeyTypes +ssh-dss' : ''}
 
@@ -342,22 +351,24 @@ Host *
     # Parameters::
     # * *host* (String): The host or IP
     def ensure_host_key(host)
-      # Get the host key
-      exit_status, stdout, _stderr = @cmd_runner.run_cmd "ssh-keyscan #{host}", timeout: TIMEOUT_HOST_KEYS, log_to_stdout: log_debug?, no_exception: true
-      if exit_status == 0
-        @platforms_ssh_dir_semaphore.synchronize do
-          known_hosts_file = "#{@platforms_ssh_dir}/known_hosts"
-          # Remove the previous eventually
-          @cmd_runner.run_cmd "ssh-keygen -R #{host} -f #{known_hosts_file}", timeout: TIMEOUT_HOST_KEYS, log_to_stdout: log_debug?
-          # Add the new one
-          host_key = stdout.strip
-          log_debug "Add new key for #{host} in #{known_hosts_file}: #{host_key}"
-          File.open(known_hosts_file, 'a') do |file|
-            file.puts host_key
+      if @strict_host_key_checking
+        # Get the host key
+        exit_status, stdout, _stderr = @cmd_runner.run_cmd "ssh-keyscan #{host}", timeout: TIMEOUT_HOST_KEYS, log_to_stdout: log_debug?, no_exception: true
+        if exit_status == 0
+          @platforms_ssh_dir_semaphore.synchronize do
+            known_hosts_file = "#{@platforms_ssh_dir}/known_hosts"
+            # Remove the previous eventually
+            @cmd_runner.run_cmd "ssh-keygen -R #{host} -f #{known_hosts_file}", timeout: TIMEOUT_HOST_KEYS, log_to_stdout: log_debug?
+            # Add the new one
+            host_key = stdout.strip
+            log_debug "Add new key for #{host} in #{known_hosts_file}: #{host_key}"
+            File.open(known_hosts_file, 'a') do |file|
+              file.puts host_key
+            end
           end
+        else
+          log_warn "Unable to get host key for #{host}. Ignoring it. Accessing #{host} might require manual acceptance of its host key."
         end
-      else
-        log_warn "Unable to get host key for #{host}. Ignoring it. Accessing #{host} might require manual acceptance of its host key."
       end
     end
 
@@ -383,24 +394,28 @@ Host *
               unless @nodes_ssh_urls.key?(node)
                 ensure_node_host_key(node)
                 ssh_url = "#{@ssh_user_name}@hpc.#{node}"
-                # Thanks to the ControlMaster option, connections are reused. So no problem to have several scp and ssh commands later.
-                log_debug "[ ControlMaster - #{node} ] - Starting ControlMaster for connection on #{ssh_url}..."
-                real_ip, _gateway, _gateway_user = connection_info_for(node, @nodes_handler.private_ip_for(node))
-                control_path_file = "/tmp/ssh_executor_mux_#{real_ip}_22_#{@ssh_user_name}"
-                if File.exist?(control_path_file)
-                  log_warn "Removing stale SSH control file #{control_path_file}"
-                  File.unlink control_path_file
-                end
-                exit_status, _stdout, _stderr = @cmd_runner.run_cmd "#{ssh_exec}#{@passwords.key?(node) ? '' : ' -o BatchMode=yes'} -o ControlMaster=yes -o ControlPersist=yes #{ssh_url} true", no_exception: no_exception, timeout: timeout
-                # Uncomment if you want to test the connection
-                # @cmd_runner.run_cmd "#{ssh_exec} -O check #{ssh_url}", log_to_stdout: false, timeout: timeout
-                if exit_status == 0
-                  # Connection ok
-                  created_ssh_urls[node] = ssh_url
-                  log_debug "[ ControlMaster - #{node} ] - ControlMaster started for connection on #{ssh_url}"
-                  @nodes_ssh_urls[node] = ssh_url
+                if @use_control_master
+                  # Thanks to the ControlMaster option, connections are reused. So no problem to have several scp and ssh commands later.
+                  log_debug "[ ControlMaster - #{node} ] - Starting ControlMaster for connection on #{ssh_url}..."
+                  real_ip, _gateway, _gateway_user = connection_info_for(node, @nodes_handler.private_ip_for(node))
+                  control_path_file = "/tmp/ssh_executor_mux_#{real_ip}_22_#{@ssh_user_name}"
+                  if File.exist?(control_path_file)
+                    log_warn "Removing stale SSH control file #{control_path_file}"
+                    File.unlink control_path_file
+                  end
+                  exit_status, _stdout, _stderr = @cmd_runner.run_cmd "#{ssh_exec}#{@passwords.key?(node) ? '' : ' -o BatchMode=yes'} -o ControlMaster=yes -o ControlPersist=yes #{ssh_url} true", no_exception: no_exception, timeout: timeout
+                  # Uncomment if you want to test the connection
+                  # @cmd_runner.run_cmd "#{ssh_exec} -O check #{ssh_url}", log_to_stdout: false, timeout: timeout
+                  if exit_status == 0
+                    # Connection ok
+                    created_ssh_urls[node] = ssh_url
+                    log_debug "[ ControlMaster - #{node} ] - ControlMaster started for connection on #{ssh_url}"
+                    @nodes_ssh_urls[node] = ssh_url
+                  else
+                    log_error "[ ControlMaster - #{node} ] - ControlMaster could not be started for connection on #{ssh_url}"
+                  end
                 else
-                  log_error "[ ControlMaster - #{node} ] - ControlMaster could not be started for connection on #{ssh_url}"
+                  @nodes_ssh_urls[node] = ssh_url
                 end
               end
             end
@@ -411,7 +426,8 @@ Host *
             created_ssh_urls.each do |node, ssh_url|
               log_debug "[ ControlMaster - #{node} ] - Stopping ControlMaster for connection on #{ssh_url}..."
               # Dumb verbose ssh! Tricky trick to just silence what is useless.
-              @cmd_runner.run_cmd "#{ssh_exec} -O exit #{ssh_url} 2>&1 | grep -v 'Exit request sent.'", expected_code: 1, timeout: timeout
+              # Don't fail if the connection close fails (but still log the error), as it can be seen as only a warning: it means the connection was closed anyway.
+              @cmd_runner.run_cmd "#{ssh_exec} -O exit #{ssh_url} 2>&1 | grep -v 'Exit request sent.'", expected_code: 1, timeout: timeout, no_exception: true
               log_debug "[ ControlMaster - #{node} ] - ControlMaster stopped for connection on #{ssh_url}"
               # Uncomment if you want to test the connection
               # @cmd_runner.run_cmd "#{ssh_exec} -O check #{ssh_url}", log_to_stdout: false, expected_code: 255, timeout: timeout
@@ -450,7 +466,9 @@ Host *
     # * *actions* (Array<[Symbol, Object]>): Ordered list of actions to perform. Each action is identified by an identifier (Symbol) and has associated data.
     #   1 action can contain several keys (action types), that will be performed in the order of the keys population in the Hash. Here are possible action types:
     #   * *local_bash* (String): Bash commands to be executed locally (not on the node)
-    #   * *ruby* (Proc): Ruby code to be executed locally (not on the node)
+    #   * *ruby* (Proc): Ruby code to be executed locally (not on the node):
+    #     * *stdout* (IO): Stream in which stdout of this action can be completed
+    #     * *stderr* (IO): Stream in which stderr of this action can be completed
     #   * *scp* (Hash<String or Symbol, String or Object>): Set of couples source => destination_dir to copy files or directories from the local file system to the remote file system. Additional options can be provided using symbols:
     #     * *sudo* (Boolean): Do we use sudo to make the copy?
     #     * *owner* (String): Owner to use for files
@@ -492,11 +510,10 @@ Host *
           when :ruby
             log_debug "[#{node}] - Execute local Ruby commands \"#{action_info}\"..."
             # TODO: Handle timeout without using Timeout which is harmful when dealing with SSH connections and multithread.
-            # TODO: Check if there is a need to handle stdout/stderr
             if @dry_run
               log_debug "[#{node}] - Won't execute Ruby code in dry_run mode."
             else
-              action_info.call
+              action_info.call action_stdout, action_stderr
             end
           when :scp
             sudo = action_info.delete(:sudo)
@@ -553,6 +570,9 @@ Host *
           stderr.concat action_stderr
           remaining_timeout -= Time.now - start_time unless remaining_timeout.nil?
         end
+      rescue CmdRunner::UnexpectedExitCodeError
+        # Error has already been logged in stderr
+        exit_status = :failed_command
       rescue
         log_error "Uncaught exception while executing actions on #{node}: #{$!}\n#{$!.backtrace.join("\n")}"
         stderr.concat "#{$!}\n"
